# docs/next-phase/openapi-v1.json
{
  "components": {
    "schemas": {
...
      "AcceptRequest": {
        "properties": {
          "business_request": {
            "additionalProperties": true,
            "title": "Business Request",
            "type": "object"
          },
          "clarification_note": {
            "default": "",
            "maxLength": 1000,
            "title": "Clarification Note",
            "type": "string"
          },
          "clarified_objective": {
            "maxLength": 2000,
            "minLength": 3,
            "title": "Clarified Objective",
            "type": "string"
          },
          "proposal_trace_id": {
            "format": "uuid",
            "title": "Proposal Trace Id",
            "type": "string"
          },
          "resolved_issue_codes": {
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "title": "Resolved Issue Codes",
            "type": "array"
          }
        },
        "required": [
          "proposal_trace_id",
          "business_request",
          "clarified_objective"
        ],
        "title": "AcceptRequest",
        "type": "object"
      },
      "AdjudicateRequest": {
        "properties": {
          "candidate_digest": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Candidate Digest",
            "type": "string"
          },
          "evidence": {
            "additionalProperties": true,
            "title": "Evidence",
            "type": "object"
          },
          "issue_code": {
            "maxLength": 64,
            "minLength": 3,
            "title": "Issue Code",
            "type": "string"
          },
          "reason": {
            "maxLength": 1000,
            "minLength": 12,
            "title": "Reason",
            "type": "string"
          },
          "root_id": {
            "format": "uuid",
            "title": "Root Id",
            "type": "string"
          }
        },
        "required": [
          "root_id",
          "candidate_digest",
          "issue_code",
          "evidence",
          "reason"
        ],
        "title": "AdjudicateRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PlanProposal": {
        "additionalProperties": false,
        "properties": {
          "candidate_actions": {
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "title": "Candidate Actions",
            "type": "array"
          },
          "entity_references": {
            "additionalProperties": {
              "type": "string"
            },
            "maxProperties": 20,
            "title": "Entity References",
            "type": "object"
          },
          "objective": {
            "maxLength": 2000,
            "minLength": 3,
            "title": "Objective",
            "type": "string"
          },
          "requested_parameters": {
            "additionalProperties": {
              "type": "string"
            },
            "maxProperties": 20,
            "title": "Requested Parameters",
            "type": "object"
          },
          "requested_workflow": {
            "maxLength": 64,
            "minLength": 3,
            "title": "Requested Workflow",
            "type": "string"
          },
          "unresolved_questions": {
            "items": {
              "type": "string"
            },
            "maxItems": 10,
            "title": "Unresolved Questions",
            "type": "array"
          }
        },
        "required": [
          "objective",
          "requested_workflow"
        ],
        "title": "PlanProposal",
        "type": "object"
      },
      "ProposeRequest": {
        "properties": {
          "context": {
            "additionalProperties": true,
            "title": "Context",
            "type": "object"
          },
          "intent": {
            "maxLength": 2000,
            "minLength": 3,
            "title": "Intent",
            "type": "string"
          }
        },
        "required": [
          "intent"
        ],
        "title": "ProposeRequest",
        "type": "object"
      },
...
      "ReviewRequest": {
        "properties": {
          "original_intent": {
            "anyOf": [
              {
                "maxLength": 2000,
                "minLength": 3,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Original Intent"
          },
          "proposal": {
            "$ref": "#/components/schemas/PlanProposal"
          },
          "proposal_trace_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proposal Trace Id"
          }
        },
        "required": [
          "proposal"
        ],
        "title": "ReviewRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "PACT - Protocol for Agentic Commit Transactions",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/planner/accept": {
      "post": {
        "operationId": "accept_api_v1_planner_accept_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-PACT-Request-ID",
            "required": true,
            "schema": {
              "title": "X-Pact-Request-Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Accept Api V1 Planner Accept Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Accept",
        "tags": [
          "planner"
        ]
      }
    },
    "/api/v1/planner/adjudicate": {
      "post": {
        "description": "Dismiss one advisory judge concern. This does not approve, freeze, or dispatch.",
        "operationId": "adjudicate_api_v1_planner_adjudicate_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-PACT-Request-ID",
            "required": true,
            "schema": {
              "title": "X-Pact-Request-Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdjudicateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Adjudicate Api V1 Planner Adjudicate Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Adjudicate",
        "tags": [
          "planner"
        ]
      }
    },
    "/api/v1/planner/assemble/{root_id}": {
      "post": {
        "operationId": "assemble_api_v1_planner_assemble__root_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "root_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Root Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Assemble Api V1 Planner Assemble  Root Id  Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Assemble",
        "tags": [
          "planner"
        ]
      }
    },
    "/api/v1/planner/explain/{tx_id}": {
      "get": {
        "operationId": "explain_api_v1_planner_explain__tx_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "tx_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Tx Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Explain Api V1 Planner Explain  Tx Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Explain",
        "tags": [
          "planner"
        ]
      }
    },
    "/api/v1/planner/propose": {
      "post": {
        "operationId": "propose_api_v1_planner_propose_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProposeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Propose Api V1 Planner Propose Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Propose",
        "tags": [
          "planner"
        ]
      }
    },
    "/api/v1/planner/review": {
      "post": {
        "operationId": "review_api_v1_planner_review_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Review Api V1 Planner Review Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Review",
        "tags": [
          "planner"
        ]
      }
    },
...
  }
}